+++ NEWS.md
@@ -1,3 +1,11 @@
+# v1.3.2
+## Adjustments
+* Exclude TTP peptide from the IgG1 quantitation.
+
+## Bug fixes
+* Minor bug fixes in spectra curation UI
+
+
 # v1.3.1
 ## Adjustments
 * Add a line of equality to the peptide correlation plots in the “IgG1 quantitation” tab.